Cleaned up the code and fixed some bugs
